moved moviedata get for movielist.js into app.js
