+++ client/src/Movies/Movie.js
@@ -1,24 +1,33 @@
 import React, { useState, useEffect } from 'react';
+import {useParams} from 'react-router-dom';
 import axios from 'axios';
 
-const Movie = (props) => {
+const Movie = ({movies}) => {
+  // Setting up state for individual movie
   const [movie, setMovie] = useState();
  
+  // Finding correct movieID 
+  const movieID = useParams();
+  console.log("here is movieID", movieID);
+  const aMovie = movies.find(item => movieID === `${item.id}`);
+  console.log("here is aMovie", aMovie);
+
+  // Getting data for individual movie
   useEffect(() => {
-    const id = 1;
     // change ^^^ that line and grab the id from the URL
     // You will NEED to add a dependency array to this effect hook
 
        axios
-        .get(`http://localhost:5000/api/movies/${id}`)
+        .get(`http://localhost:5000/api/movies/${aMovie.id}`)
         .then(response => {
+          console.log(response);
           setMovie(response.data);
         })
         .catch(error => {
           console.error(error);
         });
 
-  },[]);
+  },[aMovie]);
   
   // Uncomment this only when you have moved on to the stretch goals
   // const saveMovie = () => {
