+++ client/src/Movies/MovieList.js
@@ -1,27 +1,14 @@
-import React, { useState, useEffect } from 'react';
-import axios from 'axios';
+import React from 'react';
+import {Link} from 'react-router-dom';
 
-const MovieList = props => {
-  const [movies, setMovies] = useState([])
-  useEffect(() => {
-    const getMovies = () => {
-      axios
-        .get('http://localhost:5000/api/movies')
-        .then(response => {
-          setMovies(response.data);
-        })
-        .catch(error => {
-          console.error('Server Error', error);
-        });
-    }
-    
-    getMovies();
-  }, []);
-  
+
+const MovieList = ({movies}) => {
   return (
     <div className="movie-list">
       {movies.map(movie => (
-        <MovieDetails key={movie.id} movie={movie} />
+        <Link key={movie.id} to={`/movie/${movie.id}`}>
+          <MovieDetails movie={movie} />
+        </Link>
       ))}
     </div>
   );
